Various AUFN-related changes for Epoxy

+++ etc/kayobe/environments/aufn-ceph/a-universe-from-nothing.sh
@@ -14,16 +14,23 @@ KAYOBE_BRANCH=stackhpc/2025.1
 KAYOBE_CONFIG_BRANCH=stackhpc/2025.1
 KAYOBE_ENVIRONMENT=aufn-ceph
 
+if [[ ! -f $BASE_PATH/vault-pw ]]; then
+    echo "Vault password file not found at $BASE_PATH/vault-pw"
+    exit 1
+fi
+
 # Install git and tmux.
 if $(which dnf 2>/dev/null >/dev/null); then
     sudo dnf -y install git tmux
 else
     sudo apt update
-    sudo apt -y install git tmux gcc libffi-dev python3-dev python-is-python3
+    sudo apt -y install git tmux gcc libffi-dev python3-dev python-is-python3 python3-pip python3.12-venv
 fi
 
+export KAYOBE_VAULT_PASSWORD=$(cat $BASE_PATH/vault-pw)
+
 # Disable the firewall.
-sudo systemctl is-enabled firewalld && sudo systemctl stop firewalld && sudo systemctl disable firewalld
+sudo systemctl is-enabled firewalld && sudo systemctl stop firewalld && sudo systemctl disable firewalld || true
 
 # Disable SELinux both immediately and permanently.
 if $(which setenforce 2>/dev/null >/dev/null); then
@@ -32,7 +39,7 @@ if $(which setenforce 2>/dev/null >/dev/null); then
 fi
 
 # Prevent sudo from performing DNS queries.
-echo 'Defaults	!fqdn' | sudo tee /etc/sudoers.d/no-fqdn
+echo 'Defaults !fqdn' | sudo tee /etc/sudoers.d/no-fqdn
 
 # Clone repositories
 cd $BASE_PATH
@@ -47,7 +54,7 @@ popd
 mkdir -p venvs
 pushd venvs
 if [[ ! -d kayobe ]]; then
-    python3 -m venv kayobe
+    python3.12 -m venv kayobe
 fi
 # NOTE: Virtualenv's activate and deactivate scripts reference an
 # unbound variable.
@@ -55,7 +62,7 @@ set +u
 source kayobe/bin/activate
 set -u
 pip install -U pip
-pip install ../src/kayobe
+pip install -r ../src/kayobe-config/requirements.txt
 popd
 
 # Activate environment
